Development

+++ src/system/ScriptableSystem.h
@@ -18,7 +18,7 @@ class ScriptableSystem : public System
     public:
     ScriptableSystem(){};
 
-    void Update(double seconds_elapsed) override;
+    void Update(float seconds_elapsed) override;
     void LoadScript(lua_State *L, const std::string &script_path, const std::string &system_name);
     void OnEvent(Event& e) override;
     std::list<Subscription> GetSubscriptions() override;
